--- a/ext/zig_rb/build.zig
+++ b/ext/zig_rb/build.zig
@@ -1,6 +1,6 @@
 const std = @import("std");
 
-pub fn build(b: *std.build.Builder) void {
+pub fn build(b: *std.Build) void {
     // Standard release options allow the person running `zig build` to select
     // between Debug, ReleaseSafe, ReleaseFast, and ReleaseSmall.
     const target = b.standardTargetOptions(.{});
@@ -9,11 +9,11 @@ pub fn build(b: *std.build.Builder) void {
     const lib = b.addSharedLibrary(.{ .name = "zig_rb", .root_source_file = .{ .path = "src/main.zig" }, .version = .{ .major = 0, .minor = 0, .patch = 1 }, .optimize = optimize, .target = target });
 
     // Ruby stuff
-    var ruby_libdir = std.os.getenv("RUBY_LIBDIR") orelse "";
+    const ruby_libdir = std.posix.getenv("RUBY_LIBDIR") orelse "";
     lib.addLibraryPath(.{ .path = ruby_libdir });
-    var ruby_hdrdir = std.os.getenv("RUBY_HDRDIR") orelse "";
+    const ruby_hdrdir = std.posix.getenv("RUBY_HDRDIR") orelse "";
     lib.addIncludePath(.{ .path = ruby_hdrdir });
-    var ruby_archhdrdir = std.os.getenv("RUBY_ARCHHDRDIR") orelse "";
+    const ruby_archhdrdir = std.posix.getenv("RUBY_ARCHHDRDIR") orelse "";
     lib.addIncludePath(.{ .path = ruby_archhdrdir });
 
     lib.linkSystemLibrary("c");