@katafrakt 2025-06-03
Hey! Sorry for a delayed answer. This is a great follow up to my post.
I finally had a chance to try it out myself and it works great :heart: Thanks a lot!